+++ src/openapi/streaming/streaming.ts
@@ -63,6 +63,7 @@ type EmittedEvents = {
     [connectionConstants.EVENT_STREAMING_FAILED]: () => void;
     [connectionConstants.EVENT_CONNECTION_SLOW]: () => void;
     [connectionConstants.EVENT_DISCONNECT_REQUESTED]: () => void;
+    [connectionConstants.EVENT_MULTIPLE_ORPHANS_FOUND]: () => void;
 };
 
 /**
@@ -89,6 +90,12 @@ class Streaming extends MicroEmitter<EmittedEvents> {
      * Event that occurs when server sends _disconnect control message.
      */
     EVENT_DISCONNECT_REQUESTED = connectionConstants.EVENT_DISCONNECT_REQUESTED;
+
+    /**
+     * Event that occurs when we detect a problem with multiple orphans found
+     */
+    EVENT_MULTIPLE_ORPHANS_FOUND =
+        connectionConstants.EVENT_MULTIPLE_ORPHANS_FOUND;
 
     /**
      * Streaming has been created but has not yet started the connection.
@@ -124,6 +131,7 @@ class Streaming extends MicroEmitter<EmittedEvents> {
         connectionConstants.READABLE_CONNECTION_STATE_MAP;
 
     retryCount = 0;
+    latestActivity = 0;
     connectionState: types.ConnectionState | null =
         this.CONNECTION_STATE_INITIALIZING;
     baseUrl: string;
@@ -133,6 +141,8 @@ class Streaming extends MicroEmitter<EmittedEvents> {
     isReset = false;
     paused = false;
     orphanFinder: StreamingOrphanFinder;
+    private orphanEvents: Array<{ datetime: number; servicePath: string }> = [];
+    private multipleOrphanDetectorTimeoutId: null | number = null;
     connection!: Connection;
     connectionOptions: types.ConnectionOptions = {
         waitForPageLoad: false,
@@ -149,6 +159,7 @@ class Streaming extends MicroEmitter<EmittedEvents> {
     retryDelayLevels?: types.RetryDelayLevel[];
     reconnectTimer?: number;
     disposed = false;
+    private heartBeatLog: Array<[number, ReadonlyArray<string>]> = [];
 
     /**
      * @param transport - The transport to use for subscribing/unsubscribing.
@@ -487,6 +498,7 @@ class Streaming extends MicroEmitter<EmittedEvents> {
 
     private processUpdate(update: types.StreamingMessage) {
         try {
+            this.latestActivity = Date.now();
             if (update.ReferenceId[0] === OPENAPI_CONTROL_MESSAGE_PREFIX) {
                 this.handleControlMessage(update as types.ControlMessage);
             } else {
@@ -625,6 +637,15 @@ class Streaming extends MicroEmitter<EmittedEvents> {
     private handleControlMessageFireHeartbeats(
         heartbeatList: types.Heartbeats[],
     ) {
+        this.heartBeatLog = this.heartBeatLog.filter(
+            ([time]) => time > Date.now() - 70 * 1000,
+        );
+        this.heartBeatLog.push([
+            Date.now(),
+            heartbeatList.map(
+                ({ OriginatingReferenceId }) => OriginatingReferenceId,
+            ),
+        ]);
         log.debug(LOG_AREA, 'heartbeats received', { heartbeatList });
 
         for (let i = 0; i < heartbeatList.length; i++) {
@@ -664,12 +685,16 @@ class Streaming extends MicroEmitter<EmittedEvents> {
         referenceIdList: string[] | null,
     ) {
         if (!referenceIdList || !referenceIdList.length) {
-            log.debug(LOG_AREA, 'Resetting all subscriptions');
+            log.debug(LOG_AREA, 'Resetting all subscriptions', {
+                persist: true,
+            });
             this.resetSubscriptions(this.subscriptions.slice(0));
             return;
         }
 
-        log.debug(LOG_AREA, 'Resetting subscriptions', referenceIdList);
+        log.debug(LOG_AREA, 'Resetting subscriptions', referenceIdList, {
+            persist: true,
+        });
 
         const subscriptionsToReset = [];
         for (let i = 0; i < referenceIdList.length; i++) {
@@ -726,6 +751,10 @@ class Streaming extends MicroEmitter<EmittedEvents> {
             },
         );
 
+        this.reconnect();
+    }
+
+    reconnect() {
         this.isReset = true;
 
         // tell all subscriptions not to do anything
@@ -764,16 +793,108 @@ class Streaming extends MicroEmitter<EmittedEvents> {
         this.orphanFinder.update();
     }
 
+    private detectMultipleOrphans() {
+        if (this.multipleOrphanDetectorTimeoutId) {
+            return;
+        }
+
+        this.multipleOrphanDetectorTimeoutId = setTimeout(() => {
+            this.multipleOrphanDetectorTimeoutId = null;
+
+            const orphansByServicePath: Record<
+                string,
+                { min: number; max: number }
+            > = {};
+            for (const orphanEvent of this.orphanEvents) {
+                const orphanSpInfo =
+                    orphansByServicePath[orphanEvent.servicePath];
+                if (orphanSpInfo) {
+                    orphanSpInfo.min = Math.min(
+                        orphanSpInfo.min,
+                        orphanEvent.datetime,
+                    );
+                    orphanSpInfo.max = Math.max(
+                        orphanSpInfo.max,
+                        orphanEvent.datetime,
+                    );
+                } else {
+                    orphansByServicePath[orphanEvent.servicePath] = {
+                        min: orphanEvent.datetime,
+                        max: orphanEvent.datetime,
+                    };
+                }
+            }
+
+            let servicePathFailures = 0;
+            for (const servicePath of Object.keys(orphansByServicePath)) {
+                if (
+                    orphansByServicePath[servicePath].max -
+                        orphansByServicePath[servicePath].min >
+                    20 * 1000
+                ) {
+                    servicePathFailures++;
+                }
+            }
+
+            // multiple service paths have failed multiple times, more than 20 seconds apart, within the same minute
+            if (servicePathFailures >= 2) {
+                // debugging information...
+                const activities = this.subscriptions.map((sub) => ({
+                    latestActivity: sub.latestActivity,
+                    heartBeatLog: this.heartBeatLog,
+                    servicePath: sub.servicePath,
+                    url: sub.url,
+                    currentState: sub.currentState,
+                    inactivityTimeout: sub.inactivityTimeout,
+                    referenceId: sub.referenceId,
+                }));
+                log.error(
+                    LOG_AREA,
+                    'Detected multiple service paths hitting orphans, multiple times',
+                    {
+                        now: Date.now(),
+                        orphanEvents: this.orphanEvents,
+                        activities,
+                        latestActivity: this.latestActivity,
+                    },
+                );
+                this.trigger(this.EVENT_MULTIPLE_ORPHANS_FOUND);
+            }
+        });
+    }
+
     /**
      * Called when an orphan is found - resets that subscription
      * @param subscription - subscription
      */
     private onOrphanFound(subscription: Subscription) {
-        log.info(LOG_AREA, 'Subscription has become orphaned - resetting', {
-            referenceId: subscription.referenceId,
-            streamingContextId: subscription.streamingContextId,
-            servicePath: subscription.servicePath,
-        });
+        try {
+            log.info(LOG_AREA, 'Subscription has become orphaned - resetting', {
+                referenceId: subscription.referenceId,
+                streamingContextId: subscription.streamingContextId,
+                servicePath: subscription.servicePath,
+            });
+
+            const now = Date.now();
+            // most subscriptions time out after 60 seconds. So we use 70s to give us 10s leniency to detect multiple unsubscribes
+            const ignoreBefore = now - 70 * 1000;
+            this.orphanEvents = this.orphanEvents.filter(
+                (orphanEvent) => orphanEvent.datetime > ignoreBefore,
+            );
+            this.orphanEvents.push({
+                datetime: now,
+                servicePath: subscription.servicePath,
+            });
+
+            this.detectMultipleOrphans();
+        } catch (e) {
+            log.error(
+                LOG_AREA,
+                'Exception thrown when trying to work out multiple orphans found',
+                e,
+            );
+        }
+
         this.connection.onOrphanFound();
         subscription.reset();
     }